DEP: deprecate component.py to avoid circular imports in device.py
