+++ docs/source/conf.py
@@ -31,7 +31,7 @@ module_path = os.path.join(os.path.dirname(os.path.abspath(__file__)),'../../')
 sys.path.insert(0, module_path)
 
 import pcdsdevices  # isort: skip
-import pcdsdevices.component  # isort: skip
+import pcdsdevices.device  # isort: skip
 
 
 # -- General configuration ------------------------------------------------
