+++ pcdsdevices/device.py
@@ -1,6 +1,144 @@
-from ophyd.device import Device
+from ophyd.device import Component, Device
 
-from pcdsdevices.component import InterfaceComponent
+
+class UnrelatedComponent(Component):
+    """
+    Components that are unrelated to the main ``prefix`` attribute.
+
+    These components each need to have a full prefix specified rather than
+    the normal ``suffix`` extension or formatting. This is specified in the
+    device's ``**kwargs``, e.g. ``subdevice_prefix='SOME:PV'``.
+
+    The `collect_prefixes` class method is used to gather these special
+    prefixes in a convenient one-liner rather than needing to sift through
+    formatted components.
+
+    Examples
+    --------
+
+    A simple usage example::
+
+        from pcdsdevices.device import UnrelatedComponent as UCpt
+
+        class MyDevice(Device):
+            motor1 = UCpt(EpicsMotor)
+            motor2 = UCpt(EpicsMotor)
+
+            def __init__(self, prefix, **kwargs):
+                UCpt.collect_prefixes(self, kwargs)
+                super().__init__(prefix, **kwargs)
+    """
+
+    @classmethod
+    def collect_prefixes(cls, device, kwargs):
+        """
+        Gather all the special prefixes from a device's ``**kwargs``.
+
+        This must be called once during the ``__init__`` of a device with
+        UnrelatedComponent instances.
+
+        Parameters
+        ----------
+        device : ~ophyd.device.Device
+            The device to gather prefixes for. Typically this is just ``self``.
+
+        kwargs : dict
+            The kwargs dictionary with extra prefixes defined.
+        """
+
+        device.unrelated_prefixes = {}
+        for key, value in list(kwargs.items()):
+            if key.endswith('_prefix'):
+                device.unrelated_prefixes[key] = value
+                kwargs.pop(key)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # Needs to be non-None or it gets ignored
+        self.suffix = ''
+        add_prefix = list(kwargs.get('add_prefix', ['suffix']))
+
+        # Include subdevice UnrelatedComponent
+        try:
+            walk_iterator = self.cls.walk_components()
+        except AttributeError:
+            # No subdevices
+            walk_iterator = ()
+
+        for cpt_walk in walk_iterator:
+            if isinstance(cpt_walk.item, UnrelatedComponent):
+                name = cpt_walk.dotted_name.replace('.', '_') + '_prefix'
+                add_prefix.append(name)
+                self.kwargs[name] = None
+
+        self.add_prefix = tuple(add_prefix)
+
+    def maybe_add_prefix(self, instance, kw, suffix):
+        if kw not in self.add_prefix:
+            return suffix
+
+        # Primary prefix for UnrelatedComponent
+        if kw == 'suffix':
+            expected_kwarg = self.attr + '_prefix'
+        # Subdevice UnrelatedComponent
+        else:
+            expected_kwarg = self.attr + '_' + kw
+        try:
+            return instance.unrelated_prefixes[expected_kwarg]
+        except KeyError:
+            raise ValueError(f'Missing {expected_kwarg} in __init__ for '
+                             f'{instance.name}.')
+
+
+class ObjectComponent(Component):
+    """
+    A component that includes an existing instantiated object into a Device.
+
+    Typically, component classes exist as a description of how to create the
+    subdevice or signal, and this continues to be the preferred way to
+    define Device classes.
+
+    This exists as a convenience tool for one-off devices that need to include
+    other pre-instantiated devices as elements.
+
+    .. note::
+
+        Since this picks an object to include at class definition time,
+        you should avoid instantiating classes with object components multiple
+        times. If you would like to create a re-usable device that includes
+        pre-instantiated devices at init time, see `InterfaceDevice`.
+
+    Parameters
+    ----------
+    obj : object
+        Any existing object. This will be included in the Device as a proper
+        component, as if it was constructed by the class itself.
+    """
+    def __init__(self, obj):
+        super().__init__(object)
+        self.obj = obj
+
+    def create_component(self, instance):
+        return self.obj
+
+
+class InterfaceComponent(Component):
+    """
+    A special component to be used in the `InterfaceDevice`.
+
+    By defining a required class here, this allows us to pass
+    pre-instantiated devices into the Device constructor and type-check
+    them approriately.
+
+    Parameters
+    ----------
+    cls : type
+        Any class definition. When we pass an object into the
+        `InterfaceDevice`, its type will be checked against this type.
+        Any subclass of the input class type will also be accepted.
+    """
+    def create_component(self, instance):
+        return instance._interface_obj[self.attr]
 
 
 class InterfaceDevice(Device):
@@ -17,8 +155,8 @@ class InterfaceDevice(Device):
     .. code-block:: python
 
         from ophyd.signal import Signal
-        from pcdsdevices.device import InterfaceDevice
-        from pcdsdevices.component import InterfaceComponent as ICpt
+        from pcdsdevices.device import (InterfaceDevice,
+                                        InterfaceComponent as ICpt)
 
         class MyDevice(InterfaceDevice):
             my_component = ICpt(Signal)
