+++ tests/conftest.py
@@ -21,7 +21,7 @@ import pcdsdevices.analog_signals
 import pcdsdevices.lens
 import pcdsdevices.lxe
 from pcdsdevices.attenuator import MAX_FILTERS, Attenuator, _att_classes
-from pcdsdevices.component import UnrelatedComponent
+from pcdsdevices.device import UnrelatedComponent
 from pcdsdevices.mv_interface import setup_preset_paths
 
 MODULE_PATH = Path(__file__).parent
